feat: matching pursuit

+++ src/measurement_matrix.rs
@@ -12,7 +12,7 @@ impl<const M: usize, const N: usize> MeasurementMatrix<M, N> {
         // TODO make this injectable
         let rng = rand::thread_rng();
         // pseudo normalization
-        let norm = (1 as f64) / (M as f64);
+        let norm = 1.0 / (M as f64);
         let mut dist =
             Bernoulli::new(0.5)
                 .unwrap()
